+++ AppLovin/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
+   * 6.13.5.0
+     * Remove MREC support from AppLovin banner adapter.
+     * This version of the adapters has been certified with AppLovin SDK 6.13.5 and MoPub SDK 5.13.1.
+
    * 6.13.4.0
      * This version of the adapters has been certified with AppLovin SDK 6.13.4 and MoPub SDK 5.13.1.